pyproject.toml
[project]
name = "scrapper-rf-cnpj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "bs4>=0.0.2",
    "duckdb>=1.3.2",
    "importlib-metadata>=8.7.0",
    "jupyter>=1.1.1",
    "lxml>=5.4.0",
    "pandas>=2.2.3",
    "polars>=1.31.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-json-logger>=3.3.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=14.1.0",
    "ruff>=0.11.11",
    "sqlalchemy>=2.0.41",
    "sqlalchemy-utils>=0.41.2",
    "tenacity>=9.1.2",
    "timy>=0.4.2",
    "tqdm>=4.67.1",
    "typing-extensions>=4.13.2",
    "tzdata>=2025.2",
    "wget>=3.2",
]

